+++ package.json
@@ -27,6 +27,7 @@
     "@types/node": "20.2.5",
     "@types/react": "18.2.9",
     "@types/react-dom": "18.2.4",
+    "@uidotdev/usehooks": "^2.2.0",
     "autoprefixer": "10.4.14",
     "axios": "^1.4.0",
     "eslint": "8.42.0",
@@ -40,13 +41,15 @@
     "react-gtm-module": "^2.0.11",
     "react-hook-form": "^7.45.2",
     "react-intersection-observer": "^9.5.2",
+    "react-player": "^2.12.0",
     "react-toastify": "^9.1.3",
     "remark": "^14.0.3",
     "slugify": "^1.6.6",
     "swiper": "^9.4.1",
     "swr": "^2.1.5",
     "tailwindcss": "3.3.2",
-    "typescript": "5.1.3"
+    "typescript": "5.1.3",
+    "usehooks-ts": "^2.9.1"
   },
   "devDependencies": {
     "@tailwindcss/typography": "^0.5.9",
