+++ src/app/api/blog/[title]/route.ts
@@ -3,7 +3,7 @@ import {NextResponse} from "next/server";
 
 export async function GET(request: Request, {params} : {params: {title: string}}): Promise<Response> {
     const title = params.title
-    console.log(title)
+
     try {
         const post =  await prisma.post.findUnique({
             where: {
