+++ src/app/api/blog/route.ts
@@ -4,7 +4,15 @@ import {prisma} from "../../../../db/prisma";
 
 export async function GET(request: Request){
     try {
-        const post = await prisma.post.findMany()
+        const post = await prisma.post.findMany({
+            select: {
+                id: true,
+                title: true,
+                image_header: true,
+                createdAt: true,
+                
+            },
+        })
         return NextResponse.json({ post })
     }
     catch (e) {
