+++ src/app/blog/advice/page.tsx
@@ -26,9 +26,9 @@ export default function Advice() {
               <div className={'flex flex-col space-y-4'}>
                   {data.post.map((posts: Post) => {
                       return (
-                          <li key={posts.id as string} className={'capitalize list-disc hover:translate-x-4 hover:border-b-2 hover:border-teal-500 transition duration-300 ease-in-out'}>
+                          <div key={posts.id as string} className={'capitalize list-disc hover:translate-x-4 hover:border-b-2 hover:border-teal-500 transition duration-300 ease-in-out'}>
                               <Link className={' '}  href={`/blog/${posts.title}`}>{posts.title}</Link>
-                          </li>
+                          </div>
                       )
                   })}
               </div>
