+++ src/app/blog/card.tsx
@@ -9,8 +9,6 @@ type CardPost = {
     key: string
 }
 export default function CardBlog(props: CardPost) {
-    const slugify = require('slugify')
-    let slugUrl = slugify(props.title, {replacement : '-'})
     return (
         <>
             <div key={props.id} className={'p-4 border rounded-md border-teal-500 space-y-2 group  hover:bg-rose-500 transition duration-300 ease-in-out hover:outline-dotted hover:outline-2 hover:outline-offset-4 hover:outline-teal-700 hover:-translate-y-2'} >
@@ -18,11 +16,11 @@ export default function CardBlog(props: CardPost) {
                     <Image className="rounded-md object-cover object-top" src={`${props.image_header}`} alt={`${props.title}`} height={300} width={300} />
                 </div>
                 <div className={'capitalize '}>
-                    <p className={''}>{props.title}</p>
+                    <p className={'truncate'}>{props.title}</p>
                     <p className={''}>{props.writer} </p>
                 </div>
                 <div>
-                    <Link className={'hover:border-b-2 hover:border-blue-300'} href={`/blog/${slugUrl}`}>
+                    <Link className={'hover:border-b-2 hover:border-blue-300'} href={`/blog/${props.title}`}>
                        Read more...
                     </Link>
                 </div>
