--- src/app/career/careerdata.ts
@@ -1,20 +1,0 @@
-export interface careers {
-    id: number;
-    title: string;
-    description: string;
-    score: number;
-    level: string;
-    url: string;
-}
-export const dataCareer = [
-    { id:1, title: 'python', description: 'Great for analysis big data and Natural Language Processing', score: 4, level: 'Intermediate', url: 'https://res.cloudinary.com/dfyahyatl/image/upload/v1690166615/python_yxpxrh.png'},
-    { id:2, title: 'Next js', description: 'Web development using React framework', score: 3, level: 'Intermediate', url: 'https://res.cloudinary.com/dfyahyatl/image/upload/v1690166549/react_axzeds.png'},
-    { id:3, title: 'Nuxt js', description: 'Front end development using Vue framework', score: 3, level: 'Intermediate',url: 'https://res.cloudinary.com/dfyahyatl/image/upload/v1690165892/nuxt_kir8le.png'},
-    { id:4, title: 'Figma', description: 'For creating website design', score: 2, level: 'Beginner',url: 'https://res.cloudinary.com/dfyahyatl/image/upload/v1690166596/figma_vlihzk.png'},
-    { id:5, title: 'Google Analytic', description: 'Improve and analyse website performance', score: 3, level: 'Intermediate',url: 'https://res.cloudinary.com/dfyahyatl/image/upload/v1690166573/google-analytic_fwsrj4.png'},
-    { id:6, title: 'RDBMS', description: 'Querying for analysis & back end database', score: 2, level: 'Beginner',url: 'https://res.cloudinary.com/dfyahyatl/image/upload/v1690169530/sql_db_rwdqcz.png'},
-    { id:7, title: 'Social Network Analysis', description: 'Social media research based on big data communication with NodeXL', score: 5, level: 'Advanced',url: 'https://res.cloudinary.com/dfyahyatl/image/upload/v1690169589/AI_Network_HD_xcyl8i.png'},
-    { id:8, title: 'Power BI', description: 'Business Intelligence for ETL & analysis', score: 3, level: 'Intermediate',url: 'https://res.cloudinary.com/dfyahyatl/image/upload/v1690166608/power_bi_eyvnn4.png'},
-    { id:9, title: 'Tableau', description: 'Business Intelligence for data analysis ', score: 3, level: 'Intermediate',url: 'https://res.cloudinary.com/dfyahyatl/image/upload/v1690166628/tableau_eq8et3.png'},
-    { id:10, title: 'Adobe', description: 'Video & Photo editor and design ', score: 3, level: 'Intermediate',url: 'https://res.cloudinary.com/dfyahyatl/image/upload/v1690166585/adobe_vqdkq1.png'},
-]
