+++ src/app/content/[id]/page.tsx
@@ -24,6 +24,7 @@ export default async function Page({ params }: { params: { id: string } }) {
         try {
             await axios.patch(`/api/content/${params.id}`,
                 {title, content, tag, author, image_content, image_header, published, writer, description})
+
             toast.success("Changed !", {
                 position: toast.POSITION.BOTTOM_CENTER, theme: "dark",  icon: "🚧"
             });
@@ -79,15 +80,15 @@ export default async function Page({ params }: { params: { id: string } }) {
                             <input id={'image_content'} defaultValue={`${data.image_content}`}  {...register("image_content")} className={'rounded-md w-42 h-42 md:62 md:h-62 lg:96 lg:96'} />
                         </div>
                         <div className={'flex flex-col gap-2 text-neutral-900'}>
-                            <label htmlFor={'image_header'}>image_header</label>
-                            <input id={'image_header'} defaultValue={`${data.image_header}`} {...register("image_header")}   className={'rounded-md w-42 h-42 md:62 md:h-62 lg:96 lg:96'} />
-                        </div>
-                        <div className={'flex flex-col gap-2 text-neutral-900'}>
                             <label htmlFor={'description'}>description</label>
                             <input id={'description'} defaultValue={`${data.description}`} {...register("description")} className={'rounded-md h-[200px]'} />
                         </div>
-                        <input className={'p-2 border-sky-500 rounded-md border'} type={"submit"}/>
+                        <div className={'flex flex-col gap-2 text-neutral-900'}>
+                            <label htmlFor={'image_header'}>image_header</label>
+                            <input id={'image_header'} defaultValue={`${data.image_header}`} {...register("image_header")}   className={'rounded-md w-42 h-42 md:62 md:h-62 lg:96 lg:96'} />
+                        </div>
 
+                        <input className={'p-2 border-sky-500 rounded-md border'} type={"submit"}/>
 
                     </form>
                 </div>
